--- test.py
@@ -1,1 +1,0 @@
-from distributaur.tests import *
